refactor: separate variable declaration entry
