+++ src/domain/mod.rs
@@ -6,4 +6,5 @@ pub mod global_variable_view_factory;
 pub mod global_variables_extractor;
 pub mod type_entry;
 pub mod type_entry_repository;
-pub mod variable_declaration_repository;
+pub mod variable_declaration_entry;
+pub mod variable_declaration_entry_repository;
