+++ src/domain/variable_declaration_entry_repository.rs
@@ -1,24 +1,24 @@
 use std::ops::{Deref, DerefMut};
 
 use super::entity_repository::Repository;
-use super::global_variable::VariableDeclarationEntry;
+use super::variable_declaration_entry::VariableDeclarationEntry;
 
-pub struct VariableDeclarationRepository(Repository<VariableDeclarationEntry>);
+pub struct VariableDeclarationEntryRepository(Repository<VariableDeclarationEntry>);
 
-impl VariableDeclarationRepository {
+impl VariableDeclarationEntryRepository {
     pub fn new() -> Self {
         Self(Repository::new())
     }
 }
 
-impl Deref for VariableDeclarationRepository {
+impl Deref for VariableDeclarationEntryRepository {
     type Target = Repository<VariableDeclarationEntry>;
     fn deref(&self) -> &Self::Target {
         &self.0
     }
 }
 
-impl DerefMut for VariableDeclarationRepository {
+impl DerefMut for VariableDeclarationEntryRepository {
     fn deref_mut(&mut self) -> &mut Self::Target {
         &mut self.0
     }
