+++ tests/domain/global_variables_extractor_test.rs
@@ -4,7 +4,8 @@ use troll::domain::global_variable::*;
 use troll::domain::global_variables_extractor::*;
 use troll::domain::type_entry::*;
 use troll::domain::type_entry_repository::TypeEntryRepository;
-use troll::domain::variable_declaration_repository::VariableDeclarationRepository;
+use troll::domain::variable_declaration_entry::*;
+use troll::domain::variable_declaration_entry_repository::VariableDeclarationEntryRepository;
 use troll::library::dwarf::{DwarfInfo, DwarfInfoBuilder, DwarfTag, Location, Offset};
 
 fn init() {
@@ -20,7 +21,7 @@ fn extract_test(
     init();
 
     let mut type_entry_repository = TypeEntryRepository::new();
-    let mut variable_declaration_repository = VariableDeclarationRepository::new();
+    let mut variable_declaration_repository = VariableDeclarationEntryRepository::new();
     let mut global_variables_extractor = GlobalVariablesExtractor::new(
         &mut type_entry_repository,
         &mut variable_declaration_repository,
